refactor caldav handler and fix all-day normalization
